Add files via upload
I'm using the rigidbody velocity function to move the player, so there's no jittering or anything when you collide with the walls. There's jittering with addForce and transform. Translate ignores physics. The player's rigidbody is dynamic as opposed to kinetic, so collisions are functional. I have also constrained the player's rotation. So far, the player is incapable of rotating. I've also parented the player to the camera. Since the player cannot rotate, the camera's just fine. I also have a camera following script created but not written in, as preparation for a more elegant camera method.

For more elegant processing, the walls were removed and only colliders are used for now. This means we can sprite all the terrain, and then make collision boxes for it.